add eclipse formatter settings and VS Code settings
This file provides explicit formatting settings that can be used by VS Code or other formatters. It's currently set to the defaults, but with 4 spaces per indentation level and braces on the next line after declaring a class or method. Most of your files were consistent, but some of the indentation wasn't, and this will help keep it all nice!

Open the command palette (CMD/Ctrl + Shift + P) then search for "Format Document" to format a specific file in VS Code, or change the "Format on Save" setting to automatically format when saving a file ( I highly recomend this setting!)

This commmit also sets up VS Code to use the Eclipse formatter file.